move code to Vue SFC
