+++ templates/user/dashboard/repolist.tmpl
@@ -51,148 +51,4 @@ data.canCreateOrganization = {{.SignedUser.CanCreateOrganization}}
 window.config.pageData.dashboardRepoList = data;
 </script>
 
-<div id="dashboard-repo-list" class="six wide column">
-	<repo-search></repo-search>
-</div>
-
-<template id="dashboard-repo-list-template">
-	<div>
-		<div v-if="!isOrganization" class="ui two item tabable menu">
-			<a :class="{item: true, active: tab === 'repos'}" @click="changeTab('repos')">${textRepository}</a>
-			<a :class="{item: true, active: tab === 'organizations'}" @click="changeTab('organizations')">${textOrganization}</a>
-		</div>
-		<div v-show="tab === 'repos'" class="ui tab active list dashboard-repos">
-			<h4 class="ui top attached header gt-df gt-ac">
-				<div class="gt-f1 gt-df gt-ac">
-					${textMyRepos}
-					<span class="ui grey label gt-ml-3">${reposTotalCount}</span>
-				</div>
-				<a class="tooltip" :href="subUrl + '/repo/create'" :data-content="textNewRepo" data-position="left center">
-					<svg-icon name="octicon-plus"></svg-icon>
-					<span class="sr-only">${textNewRepo}</span>
-				</a>
-			</h4>
-			<div class="ui attached segment repos-search">
-				<div class="ui fluid right action left icon input" :class="{loading: isLoading}">
-					<input @input="changeReposFilter(reposFilter)" v-model="searchQuery" ref="search" :placeholder="textSearchRepos">
-					<i class="icon gt-df gt-ac gt-jc"><svg-icon name="octicon-search" :size="16"></svg-icon></i>
-					<div class="ui dropdown icon button" :title="textFilter">
-						<i class="icon gt-df gt-ac gt-jc gt-m-0"><svg-icon name="octicon-filter" :size="16"></svg-icon></i>
-						<div class="menu">
-							<a class="item" @click="toggleArchivedFilter()">
-								<div class="ui checkbox" ref="checkboxArchivedFilter" :title="checkboxArchivedFilterTitle">
-									<!--the "hidden" is necessary to make the checkbox work without Fomantic UI js,
-											otherwise if the "input" handles click event for intermediate status, it breaks the internal state-->
-									<input type="checkbox" class="hidden" v-bind.prop="checkboxArchivedFilterProps">
-									<label>
-										<svg-icon name="octicon-archive" :size="16" class-name="gt-mr-2"></svg-icon>
-										${textShowArchived}
-									</label>
-								</div>
-							</a>
-							<a class="item" @click="togglePrivateFilter()">
-								<div class="ui checkbox" ref="checkboxPrivateFilter" :title="checkboxPrivateFilterTitle">
-									<input type="checkbox" class="hidden" v-bind.prop="checkboxPrivateFilterProps">
-									<label>
-										<svg-icon name="octicon-lock" :size="16" class-name="gt-mr-2"></svg-icon>
-										${textShowPrivate}
-									</label>
-								</div>
-							</a>
-						</div>
-					</div>
-				</div>
-				<div class="ui secondary tiny pointing borderless menu center grid repos-filter">
-					<a class="item" :class="{active: reposFilter === 'all'}" @click="changeReposFilter('all')">
-						${textAll}
-						<div v-show="reposFilter === 'all'" class="ui circular mini grey label">${repoTypeCount}</div>
-					</a>
-					<a class="item" :class="{active: reposFilter === 'sources'}" @click="changeReposFilter('sources')">
-						${textSources}
-						<div v-show="reposFilter === 'sources'" class="ui circular mini grey label">${repoTypeCount}</div>
-					</a>
-					<a class="item" :class="{active: reposFilter === 'forks'}" @click="changeReposFilter('forks')">
-						${textForks}
-						<div v-show="reposFilter === 'forks'" class="ui circular mini grey label">${repoTypeCount}</div>
-					</a>
-					<a class="item" :class="{active: reposFilter === 'mirrors'}" @click="changeReposFilter('mirrors')" v-if="isMirrorsEnabled">
-						${textMirrors}
-						<div v-show="reposFilter === 'mirrors'" class="ui circular mini grey label">${repoTypeCount}</div>
-					</a>
-					<a class="item" :class="{active: reposFilter === 'collaborative'}" @click="changeReposFilter('collaborative')">
-						${textCollaborative}
-						<div v-show="reposFilter === 'collaborative'" class="ui circular mini grey label">${repoTypeCount}</div>
-					</a>
-				</div>
-			</div>
-			<div v-if="repos.length" class="ui attached table segment gt-rounded-bottom">
-				<ul class="repo-owner-name-list">
-					<li v-for="repo in repos" :class="{'private': repo.private || repo.internal}">
-						<a class="repo-list-link gt-df gt-ac gt-sb" :href="repo.link">
-							<div class="item-name gt-df gt-ac gt-f1 gt-mr-2">
-								<component v-bind:is="repoIcon(repo)" size="16" class="gt-mr-2"></component>
-								<div class="text gt-bold truncate gt-ml-1">${repo.full_name}</div>
-								<span v-if="repo.archived">
-									<svg-icon name="octicon-archive" :size="16" class-name="gt-ml-2"></svg-icon>
-								</span>
-							</div>
-							<div class="text light grey gt-df gt-ac" v-if="isStarsEnabled">
-								${repo.stars_count}
-								<svg-icon name="octicon-star" :size="16" class-name="gt-ml-2"></svg-icon>
-							</div>
-						</a>
-					</li>
-				</ul>
-				<div v-if="showMoreReposLink" class="center gt-py-3 gt-border-secondary-top">
-					<div class="ui borderless pagination menu narrow">
-						<a class="item navigation gt-py-2" :class="{'disabled': page === 1}"
-							@click="changePage(1)" :title="textFirstPage">
-							<svg-icon name="gitea-double-chevron-left" :size="16" class-name="gt-mr-2"></svg-icon>
-						</a>
-						<a class="item navigation gt-py-2" :class="{'disabled': page === 1}"
-							@click="changePage(page - 1)" :title="textPreviousPage">
-							<svg-icon name="octicon-chevron-left" :size="16" clsas-name="gt-mr-2"></svg-icon>
-						</a>
-						<a class="active item gt-py-2">${page}</a>
-						<a class="item navigation" :class="{'disabled': page === finalPage}"
-							@click="changePage(page + 1)" :title="textNextPage">
-							<svg-icon name="octicon-chevron-right" :size="16" class-name="gt-ml-2"></svg-icon>
-						</a>
-						<a class="item navigation gt-py-2" :class="{'disabled': page === finalPage}"
-							@click="changePage(finalPage)" :title="textLastPage">
-							<svg-icon name="gitea-double-chevron-right" :size="16" class-name="gt-ml-2"></svg-icon>
-						</a>
-					</div>
-				</div>
-			</div>
-		</div>
-		<div v-if="!isOrganization" v-show="tab === 'organizations'" class="ui tab active list dashboard-orgs">
-			<h4 class="ui top attached header gt-df gt-ac">
-				<div class="gt-f1 gt-df gt-ac">
-					${textMyOrgs}
-					<span class="ui grey label gt-ml-3">${organizationsTotalCount}</span>
-				</div>
-				<a v-if="canCreateOrganization" class="tooltip" :href="subUrl + '/org/create'" :data-content="textNewOrg" data-position="left center">
-					<svg-icon name="octicon-plus"></svg-icon>
-					<span class="sr-only">${textNewOrg}</span>
-				</a>
-			</h4>
-			<div v-if="organizations.length" class="ui attached table segment gt-rounded-bottom">
-				<ul class="repo-owner-name-list">
-					<li v-for="org in organizations">
-						<a class="repo-list-link gt-df gt-ac gt-sb" :href="subUrl + '/' + encodeURIComponent(org.name)">
-							<div class="text truncate item-name gt-f1">
-								<svg-icon name="octicon-organization" :size="16" class-name="gt-mr-2"></svg-icon>
-								<strong>${org.name}</strong>
-							</div>
-							<div class="text light grey gt-df gt-ac">
-								${org.num_repos}
-								<svg-icon name="octicon-repo" :size="16" class-name="gt-ml-2 gt-mt-1"></svg-icon>
-							</div>
-						</a>
-					</li>
-				</ul>
-			</div>
-		</div>
-	</div>
-</template>
+<div id="dashboard-repo-list" class="six wide column"></div>
